aspora_poc/requirement.txt
# Core
anthropic>=0.40.0
pydantic>=2.0
python-dotenv>=1.0.0

# Retrieval
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
rank-bm25>=0.2.2

# Graph
networkx>=3.0

# Orchestration
langgraph>=0.2.0
langchain-core>=0.3.0

# API
fastapi>=0.110.0
uvicorn>=0.27.0

# UI
streamlit>=1.30.0

# Observability (optional)
langfuse>=2.0.0

# Evaluation & Testing
pytest>=8.0.0
